+++ encrypted.backend.tf
@@ -1,0 +1,22 @@
+{
+	"data": "ENC[AES256_GCM,data:w9PiQmTXW7Po8/u5Y+BEH9gdPT3UnaXeRQhPfdL9bXN/1h/sONfP1c1i0UwODCwI7fyf40eJ3ZdUUTq9NNFzfGbM2Kli4/WF+Hd/M5abeBFGk0qMcia4/XYHbYuAF3sOuV51gSsqM+5YssOz6BujwvkgBvwi4qQM4hF+tR6cGZcsgYtMa6OQSjjcdZb11cVo915BqEPwR36xCeO1UPITVGD8RX0lWgpdmMj9EAu3HlaB11DI/viD0JLk4DQ33tEYNdR+I3Rx9Ry4J6IAhfZExng1uBCdtcvDSApYSQqa282YyVQz8raMqZHSPh5xgYTjutIPpxxFd0pnCplt38lwCcuCsZsV8KKfuPfdml/8d0IyU+MOVq+DKPJivdqcHycKJMziR9rtNCVzfmGZTJTqiQqAQTJ71FF4y9khcjoqzvzeXhR1gJlMkr1Do76mHx8RT4pofiI19mzYx/dW6Z7eYyoL7NV0dnWAoltzEhxsw9mAtWlIi0Ui8pJGvWRabJ7XqhMQeg==,iv:z9jfj9extEaq2/9I4RmlBPSGKktLNmsRV8ucgjYqZF8=,tag:BvpbbbGP/I5pJXyHXafcOw==,type:str]",
+	"sops": {
+		"kms": [
+			{
+				"arn": "arn:aws:kms:ap-southeast-1:679326800503:key/3cdf98ff-1261-4c39-823e-f9feee5ad5b9",
+				"created_at": "2023-08-12T13:48:10Z",
+				"enc": "AQICAHg5TwXQ3UYJQ4LBo3tCfjVcN12M/cM0LY4wILn8Q/7a+QG85NPXt+xYqkhDf6CONsQ9AAAAfjB8BgkqhkiG9w0BBwagbzBtAgEAMGgGCSqGSIb3DQEHATAeBglghkgBZQMEAS4wEQQMsL4Gy4vIBzsXZxzMAgEQgDsQrcnNmBzPbRLKmUmlp0h90xSGsHWeYTeFlhdi0PRMuD/1YDag5nUx5a/pS+EaN4fp3LDwuVCHyEjcTw==",
+				"aws_profile": ""
+			}
+		],
+		"gcp_kms": null,
+		"azure_kv": null,
+		"hc_vault": null,
+		"age": null,
+		"lastmodified": "2023-08-12T13:48:10Z",
+		"mac": "ENC[AES256_GCM,data:KpWVFfKz99wOVbIxqc2r4ywueBiOd6/etgN8vOn4tbzIEtLO9EQPjz5YH4K6fWJxpnphpn7eX4ojO5odnEorfNoooC/tXkELCNk8yjCHZvDPICxuosiPYIacJiCG5gTmavgy5Y+g3ltyHo3bogDSNG8196AqAgbmkLPaoat3Z+o=,iv:32XI6KeEomxq4ee1J/ToQQrKfcB44WP/CETUXR9U83c=,tag:dWNO02/3T/tjTeCQcyBVSg==,type:str]",
+		"pgp": null,
+		"unencrypted_suffix": "_unencrypted",
+		"version": "3.7.3"
+	}
+}