+++ CHANGELOG.md
@@ -13,6 +13,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 # [Released]
 
+## [0.0.7] - 2023-08-12
+### Added
+- added s3, iam, vpc, r53, waf modeules
+- edits to terragrunt and terraform products
+- added encrypted backend
+
 ## [0.0.6] - 2023-08-12
 ### Added
 - added vendir